Refactor login options into shared AuthOptions component
